--- a/tests/test_single_table_blendsql.py
+++ b/tests/test_single_table_blendsql.py
@@ -592,7 +592,10 @@ def test_apply_limit_with_predicate(db, ingredients):
         """
     SELECT COUNT(DISTINCT merchant) FROM transactions WHERE amount > 1300 LIMIT 3    """
     )[0]
-    assert smoothie.meta.num_values_passed == passed_to_ingredient
+    # We say `<=` here because the ingredient operates over sets, rather than lists
+    # So this kind of screws up the `LIMIT` calculation
+    # But execution outputs should be the same (tested above)
+    assert smoothie.meta.num_values_passed <= passed_to_ingredient
 
 
 if __name__ == "__main__":