On queries like `benchmarks/national_parks/q05.sql`:

```sql
SELECT {{ImageCaption('parks::Image')}} as "Image Description" FROM parks LIMIT 1
```

We apply the `LIMIT` clause (and any other surrounding filters) prior to calling the `ImageCaption` ingredient

## Benchmark Results:
Before:
| Task           |   Average Runtime |   # Unique Queries |
|:---------------|------------------:|-------------------:|
| financials     |         0.0402491 |                  7 |
| rugby          |         0.323255  |                  4 |
| national_parks |         2.07314   |                  5 |
| 1966_nba_draft |         0.119926  |                  2 |

After:
| Task           |   Average Runtime |   # Unique Queries |
|:---------------|------------------:|-------------------:|
| financials     |          0.040797 |                  7 |
| rugby          |          0.319473 |                  4 |
| national_parks |          0.865904 |                  5 |
| 1966_nba_draft |          0.115434 |                  2 |